refactor device.ts to split up code
