+++ www/src/ts/app/nav.ts
@@ -1,6 +1,6 @@
 import { HTMLTemplateResult, html, css } from "lit";
 import { customElement, property } from "lit/decorators.js";
-import { BaseElement, defaultCss } from "../components";
+import { BaseElement, defaultCss } from "components";
 
 import "@shoelace-style/shoelace/dist/components/icon/icon.js";
 import "@shoelace-style/shoelace/dist/components/icon-button/icon-button.js";
