+++ www/src/ts/virtual_machine/device/utils.ts
@@ -1,5 +1,5 @@
 import { Connection } from "ic10emu_wasm";
-import { DeviceDBConnection } from "./device_db";
+import { DeviceDBConnection } from "../device_db";
 
 const CableNetworkTypes: readonly string[] = Object.freeze(["Power", "Data", "PowerAndData"]);
 export function connectionFromDeviceDBConnection(conn: DeviceDBConnection): Connection {
