+++ www/src/ts/virtual_machine/stack.ts
@@ -1,14 +1,14 @@
 import { html, css } from "lit";
 import { customElement } from "lit/decorators.js";
-import { BaseElement, defaultCss } from "../components";
-import { VMActiveICMixin } from "./base_device";
+import { BaseElement, defaultCss } from "components";
+import { VMActiveICMixin } from "virtual_machine/base_device";
 
 import "@shoelace-style/shoelace/dist/components/card/card.js";
 import "@shoelace-style/shoelace/dist/components/icon/icon.js";
 import "@shoelace-style/shoelace/dist/components/tooltip/tooltip.js";
 import "@shoelace-style/shoelace/dist/components/input/input.js";
 import SlInput from "@shoelace-style/shoelace/dist/components/input/input.js";
-import { displayNumber, parseNumber } from "../utils";
+import { displayNumber, parseNumber } from "utils";
 
 @customElement("vm-ic-stack")
 export class VMICStack extends VMActiveICMixin(BaseElement) {
