+++ www/src/ts/virtual_machine/ui.ts
@@ -1,6 +1,6 @@
-import { HTMLTemplateResult, html, css } from "lit";
-import { customElement, property, query, state } from "lit/decorators.js";
-import { BaseElement, defaultCss } from "../components";
+import { html, css } from "lit";
+import { customElement } from "lit/decorators.js";
+import { BaseElement, defaultCss } from "components";
 import "@shoelace-style/shoelace/dist/components/details/details.js";
 import "@shoelace-style/shoelace/dist/components/tab/tab.js";
 import "@shoelace-style/shoelace/dist/components/tab-panel/tab-panel.js";
